Added support for albumentations
I adapted the albumentations support from mmdet with the use of mask (instead of masks and bboxes). I've tested it with some extent and should be working seamingless with those albumentations that expect a semantic segmentation.

The only concern is: albumentations works with RGB images, while mmcv reads them in BGR format by default. I've added the explicit conversion at lines 1455 and 1460, but I've seen that in mmdet this is not taken in consideration. Is the conversion made somewhere else in that case?